Start bulk indexing from cincoctrl management
