+++ cincoctrl/config/settings/base.py
@@ -88,7 +88,7 @@ LOCAL_APPS = [
 # https://docs.djangoproject.com/en/dev/ref/settings/#installed-apps
 INSTALLED_APPS = DJANGO_APPS + THIRD_PARTY_APPS + LOCAL_APPS
 
-AIRFLOW_PROJECT_NAME = "cincoctrl"
+AIRFLOW_PROJECT_NAME = "cinco"
 AIRFLOW_JOB_RELATED_MODEL = "findingaids.FindingAid"
 AIRFLOW_ENV_NAME = "pad-airflow-mwaa"
 # no trailing slash on AIRFLOW_ENV_URL
